+++ ship/LambdaShuttle.cpp
@@ -2,7 +2,7 @@
 
 const std::string LambdaShuttle::MODEL = "Lambda-class shuttle";
 const double LambdaShuttle::WEIGHT = 360;
-const unsigned LambdaShuttle::SPEED = 54;
+const double LambdaShuttle::SPEED = 54;
 const double LambdaShuttle::MAX_LOAD = 80;
 unsigned LambdaShuttle::id = 0;
