+++ ship/LoadedShip.hpp
@@ -8,7 +8,7 @@ public:
     std::ostream& toStream(std::ostream& os) const override;
 
 protected:
-    LoadedShip(double weight, unsigned speed, const std::string& model, unsigned id,
+    LoadedShip(double weight, double speed, const std::string& model, unsigned id,
                double maxLoad,
                double currentLoad = 0);
