+++ ship/StarDestroyer.cpp
@@ -2,7 +2,7 @@
 
 const std::string StarDestroyer::MODEL = "Super-class Star Destroyer";
 const double StarDestroyer::WEIGHT = 9000000000;
-const unsigned StarDestroyer::SPEED = 40;
+const double StarDestroyer::SPEED = 40;
 const double StarDestroyer::MAX_LOAD = 250000;
 unsigned StarDestroyer::id = 0;
