+++ ship/TIEIN.cpp
@@ -2,7 +2,7 @@
 
 const std::string TIEIN::MODEL = "TIE/IN";
 const double TIEIN::WEIGHT = 5;
-const unsigned TIEIN::SPEED = 110;
+const double TIEIN::SPEED = 110;
 unsigned TIEIN::id = 0;
 
 TIEIN::TIEIN() : Ship(WEIGHT, SPEED, MODEL, ++id) {}
