+++ ship/TIELN.cpp
@@ -2,7 +2,7 @@
 
 const std::string TIELN::MODEL = "TIE/LN";
 const double TIELN::WEIGHT = 6;
-const unsigned TIELN::SPEED = 100;
+const double TIELN::SPEED = 100;
 unsigned TIELN::id = 0;
 
 TIELN::TIELN() : Ship(WEIGHT, SPEED, MODEL, ++id) {}
