+++ ship/TIELN.hpp
@@ -17,7 +17,7 @@ public:
 private:
     static const std::string MODEL;
     static const double WEIGHT;
-    static const unsigned SPEED;
+    static const double SPEED;
     static unsigned id;
 };
